% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp-steps.R
\name{pp_ecg_read}
\alias{pp_ecg_read}
\title{Read a WFDB record (by companion files or base) into a PhysioExperiment}
\usage{
pp_ecg_read(record)
}
\arguments{
\item{record}{Either the companion files (\code{c(".dat", ".hea")}) or a
single record base path (no extension).}
}
\value{
A \code{PhysioExperiment}.
}
\description{
Read a WFDB record (by companion files or base) into a PhysioExperiment
}
